# [multifn] Optimize non-cached effective-dispatch-value